requirements.txt
asgiref==3.8.1
black==24.8.0
boto3==1.35.10
botocore==1.35.10
Django==5.1
django-debug-toolbar==4.4.6
django-filter==24.3
django-markdownx==4.0.7
django-ses==4.1.1
django-storages==1.14.4
djangorestframework==3.15.2
jmespath==1.0.1
Markdown==3.7
pillow==10.4.0
psycopg==3.2.1
psycopg-binary==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
